integrate omni node
