+++ Cargo.toml
@@ -20,6 +20,7 @@ members = [
 async-trait = { version = "0.1.74" }
 clap        = { version = "4.5.1", default-features = false }
 color-print = { version = "0.3.5" }
+color-eyre = { version = "0.6.3", default-features = false }
 futures     = { version = "0.3.28", default-features = false }
 hex-literal = { version = "0.4.1", default-features = false }
 hex         = { version = "0.4.3", default-features = false }
@@ -114,6 +115,7 @@ substrate-state-trie-migration-rpc         = { version = "44.0.0" }
 # Polkadot
 pallet-xcm                    = { version = "20.1.2", default-features = false }
 polkadot-cli                  = { version = "25.0.0" }
+polkadot-omni-node-lib        = { version = "0.7.0" }
 polkadot-parachain-primitives = { version = "17.0.0", default-features = false }
 polkadot-primitives           = { version = "19.0.0", default-features = false }
 polkadot-runtime-common       = { version = "20.0.0", default-features = false }
